✨ Allow demo mode on register page + logout on deletion
