+++ apps/wallet/app/module/authentication/component/Logout/index.tsx
@@ -1,76 +1,21 @@
 import { Button } from "@frak-labs/design-system/components/Button";
 import { Spinner } from "@frak-labs/design-system/components/Spinner";
-import {
-    recoveryHintStorage,
-    sessionStore,
-    trackEvent,
-} from "@frak-labs/wallet-shared";
-import { useQueryClient } from "@tanstack/react-query";
-import { useNavigate } from "@tanstack/react-router";
 import { LogOut } from "lucide-react";
-import { useState } from "react";
 import { useTranslation } from "react-i18next";
-import { panelDismissedPrefix } from "@/module/common/component/Panel";
-import { notificationAdapter } from "@/module/notification/adapter";
+import { useLogout } from "@/module/authentication/hook/useLogout";
 import * as styles from "./index.css";
-
-function cleanLocalStorage() {
-    // Clear static local storage items
-    const localStorageItems = [
-        "REACT_QUERY_OFFLINE_CACHE",
-        "frak_theme",
-        "frak_session",
-        "frak_sdkSession",
-        "frak_lastWebAuthNAction",
-        "frak_user",
-        "frak_userSetupLater",
-    ];
-    for (const item of localStorageItems) {
-        window.localStorage.removeItem(item);
-    }
-
-    // Clear all dismissed panel states
-    for (const key of Object.keys(window.localStorage)) {
-        if (key.startsWith(panelDismissedPrefix)) {
-            window.localStorage.removeItem(key);
-        }
-    }
-}
 
 /**
  * Logout from current authentication
- * @constructor
  */
 export function Logout() {
     const { t } = useTranslation();
-    const navigate = useNavigate();
-    const queryClient = useQueryClient();
-    const [isLoggingOut, setIsLoggingOut] = useState(false);
-
-    const handleLogout = async () => {
-        setIsLoggingOut(true);
-        trackEvent("logout");
-        // Unsubscribe from push notifications before clearing session (needs auth)
-        await notificationAdapter.unsubscribe().catch(() => {});
-        // Wipe the uninstall-resilient recovery hint from iCloud KV /
-        // Block Store. Without this, a fresh install would still redirect
-        // to /login because the hint persists in the user's cloud account.
-        await recoveryHintStorage.clear();
-        // Session deletion
-        sessionStore.getState().clearSession();
-        // Query cache
-        queryClient.removeQueries();
-        // Local storage cleanup
-        setTimeout(() => {
-            cleanLocalStorage();
-            navigate({ to: "/register", replace: true });
-        }, 100);
-    };
+    const { logout, isLoggingOut } = useLogout();
 
     return (
         <Button
             disabled={isLoggingOut}
-            onClick={handleLogout}
+            onClick={logout}
             icon={isLoggingOut ? <Spinner size="s" /> : <LogOut size={20} />}
             className={styles.logoutButton}
         >
