+++ packages/wallet-shared/src/i18n/locales/en/translation.json
@@ -305,7 +305,7 @@
             "title": "Try Frak in demo mode",
             "description": "Spin up a temporary wallet to explore Frak. No biometrics, no recovery — just a sandbox.",
             "button": {
-                "create": "Create your demo <strong>wallet</strong>",
+                "create": "Create your demo wallet",
                 "error": "Error during registration, please try again",
                 "inProgress": "Wallet creation in progress"
             }
